test-requirements.txt
backports.tempfile; python_version == '2.7'
tox
virtualenv
